requirements.txt
absl-py==2.1.0
accelerate==1.2.1
anyio==4.7.0
fastapi==0.85.0
google-auth==2.37.0
google-auth-oauthlib==0.4.6
h5py==3.12.1
huggingface-hub==0.27.0
Jinja2==3.1.5
keras==2.11.0
matplotlib==3.5.3
moviepy==1.0.3
numpy==1.26.4
opencv-python==4.6.0.66
pandas==2.2.3
pydantic==1.10.19
pydub==0.25.1
PyPDF2==3.0.1
python-docx==0.8.11
scikit-learn==1.3.0
scipy==1.15.0
tensorflow==2.11.0
torch==1.12.1+cu113
torchaudio==0.12.1+cu113
torchvision==0.13.1+cu113
transformers==4.47.1
uvicorn==0.18.3